Rename and move some classes
* server/service/GoDashboardCurrentStateLoader.java =>
  server/dashboard/GoDashboardCurrentStateLoader.java

* GoDashboardCacheUpdateService => GoDashboardService (to reduce
  confusion and make it a single point of access)
