+++ server/src/com/thoughtworks/go/server/dashboard/GoDashboardConfigChangeHandler.java
@@ -18,17 +18,17 @@ package com.thoughtworks.go.server.dashboard;
 
 import com.thoughtworks.go.config.CruiseConfig;
 import com.thoughtworks.go.config.PipelineConfig;
-import com.thoughtworks.go.server.service.GoDashboardCacheUpdateService;
+import com.thoughtworks.go.server.service.GoDashboardService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 /* Understands what needs to be done to keep the dashboard cache updated, when the config changes. */
 @Component
 public class GoDashboardConfigChangeHandler {
-    private GoDashboardCacheUpdateService cacheUpdateService;
+    private GoDashboardService cacheUpdateService;
 
     @Autowired
-    public GoDashboardConfigChangeHandler(GoDashboardCacheUpdateService cacheUpdateService) {
+    public GoDashboardConfigChangeHandler(GoDashboardService cacheUpdateService) {
         this.cacheUpdateService = cacheUpdateService;
     }
