+++ server/src/com/thoughtworks/go/server/dashboard/GoDashboardPipelinePauseStatusChangeHandler.java
@@ -17,20 +17,18 @@
 package com.thoughtworks.go.server.dashboard;
 
 import com.thoughtworks.go.config.CaseInsensitiveString;
-import com.thoughtworks.go.config.CruiseConfig;
-import com.thoughtworks.go.config.PipelineConfig;
 import com.thoughtworks.go.server.domain.PipelinePauseChangeListener;
-import com.thoughtworks.go.server.service.GoDashboardCacheUpdateService;
+import com.thoughtworks.go.server.service.GoDashboardService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 /* Understands what needs to be done to keep the dashboard cache updated, when a pipeline is paused or unpaused. */
 @Component
 public class GoDashboardPipelinePauseStatusChangeHandler {
-    private GoDashboardCacheUpdateService cacheUpdateService;
+    private GoDashboardService cacheUpdateService;
 
     @Autowired
-    public GoDashboardPipelinePauseStatusChangeHandler(GoDashboardCacheUpdateService cacheUpdateService) {
+    public GoDashboardPipelinePauseStatusChangeHandler(GoDashboardService cacheUpdateService) {
         this.cacheUpdateService = cacheUpdateService;
     }
