+++ server/test/unit/com/thoughtworks/go/server/dashboard/GoDashboardPipelinePauseStatusChangeHandlerTest.java
@@ -19,18 +19,17 @@ package com.thoughtworks.go.server.dashboard;
 import com.thoughtworks.go.config.CaseInsensitiveString;
 import com.thoughtworks.go.server.domain.PipelinePauseChangeListener;
 import com.thoughtworks.go.server.domain.Username;
-import com.thoughtworks.go.server.service.GoDashboardCacheUpdateService;
+import com.thoughtworks.go.server.service.GoDashboardService;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
 
-import static org.junit.Assert.*;
 import static org.mockito.Mockito.verify;
 import static org.mockito.MockitoAnnotations.initMocks;
 
 public class GoDashboardPipelinePauseStatusChangeHandlerTest {
     @Mock
-    private GoDashboardCacheUpdateService cacheUpdateService;
+    private GoDashboardService cacheUpdateService;
 
     private GoDashboardPipelinePauseStatusChangeHandler handler;
